This is good
My Day is gone and is replaced with Day Tracker since it's no longer
limited to one day of tracking! The program will keep adding it to the
same .txt and won't rewrite it every time you launch it!

Will add to Day Tracker if the file exists (in same folder), or will
create Day Tracker if it doesn't exist (and then add stuff)